Add files via upload
This is a customer order and invoice database, modeled  based on a typical workflow sequence of a physical or an online store using the popular data manipulation language, SQL. It comprises of tables, relationships and other objects. In the scripts various operations were carried out such as create, retrieve, insert, update and delete.